# avoid duplicate inplaceSortBy code