--- a/basement/Basement/Algorithm.hs
+++ b/basement/Basement/Algorithm.hs
@@ -1,5 +1,6 @@
 module Basement.Algorithm
-    ( inplaceSortBy
+    ( RandomAccess, read, write
+    , inplaceSortBy
     ) where
 
 import           GHC.Types
@@ -11,15 +12,18 @@ import           Basement.Types.OffsetSize
 import           Basement.PrimType
 import           Basement.Monad
 
-inplaceSortBy :: (PrimType ty, PrimMonad prim) 
+class RandomAccess a where
+    read  :: (PrimType ty, PrimMonad prim) => a (PrimState prim) -> (Offset ty) -> prim ty
+    write :: (PrimType ty, PrimMonad prim) => a (PrimState prim) -> (Offset ty) -> ty -> prim ()
+
+inplaceSortBy :: (PrimType ty, PrimMonad prim, RandomAccess a) 
               => (ty -> ty -> Ordering)
               -- ^ Function defining the ordering relationship
               -> (Offset ty) -- ^ Offset to first element to sort
               -> (CountOf ty) -- ^ Number of elements to sort
-              -> ((Offset ty -> prim ty), (Offset ty  -> ty -> prim ()))
-              -- ^ Pair of read and write actions for a given offset
+              -> a (PrimState prim) -- ^ Data to be sorted
               -> prim ()
-inplaceSortBy ford start len (unsafeRead, unsafeWrite) 
+inplaceSortBy ford start len mvec
     = qsort start (start `offsetPlusE` len `offsetSub` 1)
     where
         qsort lo hi
@@ -30,9 +34,9 @@ inplaceSortBy ford start len (unsafeRead, unsafeWrite)
                 qsort (p+1) hi
         pivotStrategy (Offset low) hi@(Offset high) = do
             let mid = Offset $ (low + high) `div` 2
-            pivot <- unsafeRead mid
-            unsafeRead hi >>= unsafeWrite mid
-            unsafeWrite hi pivot -- move pivot @ pivotpos := hi
+            pivot <- read mvec mid
+            read mvec hi >>= write mvec mid
+            write mvec hi pivot -- move pivot @ pivotpos := hi
             pure pivot
         partition lo hi = do
             pivot <- pivotStrategy lo hi
@@ -40,14 +44,14 @@ inplaceSortBy ford start len (unsafeRead, unsafeWrite)
             -- INVARIANT: i & j are valid array indices; pivotpos==hi
             let go i j = do
                     -- INVARIANT: k <= pivotpos
-                    let fw k = do ak <- unsafeRead k
+                    let fw k = do ak <- read mvec k
                                   if ford ak pivot == LT 
                                     then fw (k+1)
                                     else pure (k, ak)
                     (i, ai) <- fw i -- POST: ai >= pivot
                     -- INVARIANT: k >= i
                     let bw k | k==i = pure (i, ai)
-                             | otherwise = do ak <- unsafeRead k
+                             | otherwise = do ak <- read mvec k
                                               if ford ak pivot /= LT
                                                 then bw (pred k)
                                                 else pure (k, ak)
@@ -56,14 +60,14 @@ inplaceSortBy ford start len (unsafeRead, unsafeWrite)
                     if i < j
                         then do -- (ai>=p AND aj<p) AND (i<j<pivotpos)
                             -- swap two non-pivot elements and proceed
-                            unsafeWrite i aj
-                            unsafeWrite j ai
+                            write mvec i aj
+                            write mvec j ai
                             -- POST: (ai < pivot <= aj)
                             go (i+1) (pred j)
                         else do -- ai >= pivot 
                             -- complete partitioning by swapping pivot to the center
-                            unsafeWrite hi ai 
-                            unsafeWrite i pivot
+                            write mvec hi ai 
+                            write mvec i pivot
                             pure i
             go lo hi
 {-# INLINE inplaceSortBy #-}