By passing a pair of read & write actions to `inplaceSortBy` the algorithm becomes independent of the underlying representation (Native/Foreign), while still sorting the arrays with specialized code. The benchmark 'types/ByteArray/Sort' showed the same execution time (before and after the change). Hence, we can avoid the duplication of the sort logic. I guess the same idea could be applied to other functions in the `Basement.Alg.*` module. @vincenthz What do you think about that approach?