@vincenthz 2017-08-27
I think this is only workable when the type is specialized, otherwise a quick benchmark tell me that it this approach goes slower when ghc hasn't specialized the prim function to a specific type (here with array of Word32), which are very likely the result of further boxes (tuple, function not resolved at compilation,etc):

before:

```
benchmarking types/ByteArray/Sort32/bs2000/UArray_W32
time                 984.3 μs   (980.6 μs .. 987.7 μs)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 983.3 μs   (979.4 μs .. 996.6 μs)
std dev              21.76 μs   (5.438 μs .. 54.15 μs)
variance introduced by outliers: 12% (moderately inflated)
```

after:

```
benchmarking types/ByteArray/Sort32/bs2000/UArray_W32
time                 1.105 ms   (1.100 ms .. 1.110 ms)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 1.103 ms   (1.101 ms .. 1.107 ms)
std dev              8.788 μs   (5.793 μs .. 13.49 μs)
```