@ciderale 2017-08-27
Okay, I see. You are right, there is some overhead involved. I will experiment with the  'types/ByteArray/Sort32/bs2000/UArray_W32' benchmark to see if it is possible to overcome the problem. I still hope that the code duplication can be avoided.