@ndmitchell 2017-08-27
One trick to consider is using the static argument transformation on the two input arguments, as an example:

```haskell
{-# INLINE sortBy #-}
inplaceSortBy unsafeRead unsafeWrite = myInplaceSortBy
    where
        myInplaceSortBy ford start len = ... using unsafeRead ...
```

I suggest putting the two unsafeRead/write as the first two arguments, and not having the tuple, since that will definitely improve GHC's chances of matching without duplication (I think the duplication is really yuk...)