@ciderale 2017-08-27
Thanks for the input. I was currently experimenting with a typeclass:
```
class RandomAccess a ty where
    read  :: PrimMonad prim => a ty (PrimState prim) -> (Offset ty)       -> prim ty
    write :: PrimMonad prim => a ty (PrimState prim) -> (Offset ty) -> ty -> prim ()

inplaceSortBy :: (PrimMonad prim, RandomAccess a ty) 
              => (ty -> ty -> Ordering) -> (Offset ty) -> (CountOf ty)
              -> a ty (PrimState prim) -- ^ Data to be sorted
              -> prim ()

instance PrimType ty => Algorithm.RandomAccess MutableBlock ty where
    read (MutableBlock mba) = primMbaRead mba
    write (MutableBlock mba) = primMbaWrite mba

data PtrSt ty st = PtrSt Addr#
instance PrimType ty => Algorithm.RandomAccess PtrSt ty where
    read (PtrSt addr) = PrimAddr.primRead addr
    write (PtrSt addr) = PrimAddr.primWrite addr
```
For UArray the benchmark looks good and the core output (before/after) are identical (modulo variable renaming). 

I'm a bit unsure if this is too heavy machinery, since I had to enable MultiParamTypeClasses and FlexibleContexts. Especially FlexibleContexts is something that I'm not really familiar with. So any feedback on the above (including naming) are very appreciated.

Excluding the `PrimType ty` constraint from the class definition was a deliberate choice. This allows to reuse `inplaceSortBy` for `BoxedArray` types, thus reducing one more duplicate of the sorting routine.