@ndmitchell 2017-08-27
Your typeclass extensions are positively pedestrian - on some projects I have them on by default. For reference, I would have gone with:

```
class RandomAccess container prim ty where
    read  :: container -> Offset ty   -> prim ty
    write :: container -> Offset ty -> ty -> prim ()
```

The requirement that `a` have the shape it does seems very artificial - you couldn't use it for `MutableBoolVector` on top of `Integer`, for example. You also hoist the `PrimMonad` class one level up, which seems cleaner.