@ciderale 2017-08-27
Thanks for the clarification and the cleaned-up class. I will make the necessary changes to integrate this.

Regarding your last comment about the static argument transformation. That was already done
```
inplaceSortBy ford start len (unsafeRead, unsafeWrite) 
    = qsort start (start `offsetPlusE` len `offsetSub` 1) where qsort ... use unsafeRead ...
```
so I could only try avoiding the tuple.