@ndmitchell 2017-08-27
Avoiding the tuple is a good idea. But so is moving all arguments that don't change "that much" further up in the list - e.g. `unsafeRead` never changes but `len` often does. You might also get benefits from having the two lambdas - so the first only takes `unsafeRead`/`write`, and a subsequent lambda gets the changing arguments.