@ciderale 2017-08-27
The typeclass based approach does, as far as I can tell, not incur the overhead of the previous approach based on tuples. Simply changing tuples to individual arguments didn't show any improvements. Based on the typeclass abstraction, it was also possible to change BoxedArray.sortBy to use Algorithm.inplaceSortBy. 

The typeclass definition `class RandomAccess container prim ty` (rather the one in the my earlier comment) was already helpful as `Ptr ty` (for the Foreign backend) is not parametrized by prim. 

Below are the benchmark results on Word32 data. The result for UArray and Block are comparable.
The significant gain on Array_W32 is due to the fact of using the new partitioning scheme for quicksort that was merged in PR #390.

I'm curious what you think about the proposed changes.

Benchmarks original:
```
benchmarking types/ByteArray/Sort32/2000/Array_W32
time                 1.681 ms   (1.651 ms .. 1.714 ms)
                     0.995 R²   (0.990 R² .. 0.998 R²)
mean                 1.721 ms   (1.693 ms .. 1.759 ms)
std dev              112.5 μs   (92.46 μs .. 137.1 μs)
variance introduced by outliers: 48% (moderately inflated)

benchmarking types/ByteArray/Sort32/2000/UArray_W32
time                 1.007 ms   (987.3 μs .. 1.023 ms)
                     0.997 R²   (0.995 R² .. 0.998 R²)
mean                 1.004 ms   (991.0 μs .. 1.021 ms)
std dev              54.79 μs   (43.52 μs .. 68.11 μs)
variance introduced by outliers: 44% (moderately inflated)

benchmarking types/ByteArray/Sort32/2000/Block_W32
time                 1.018 ms   (989.8 μs .. 1.051 ms)
                     0.994 R²   (0.990 R² .. 0.998 R²)
mean                 1.009 ms   (994.1 μs .. 1.028 ms)
std dev              53.42 μs   (43.56 μs .. 68.92 μs)
variance introduced by outliers: 43% (moderately inflated)
```

Benchmark after:
```
benchmarking types/ByteArray/Sort32/2000/Array_W32
time                 244.1 μs   (239.3 μs .. 250.6 μs)
                     0.992 R²   (0.983 R² .. 0.997 R²)
mean                 258.3 μs   (252.5 μs .. 270.9 μs)
std dev              29.10 μs   (14.78 μs .. 53.22 μs)
variance introduced by outliers: 83% (severely inflated)

benchmarking types/ByteArray/Sort32/2000/UArray_W32
time                 1.005 ms   (993.6 μs .. 1.017 ms)
                     0.998 R²   (0.998 R² .. 0.999 R²)
mean                 1.003 ms   (992.2 μs .. 1.018 ms)
std dev              46.29 μs   (34.90 μs .. 62.15 μs)
variance introduced by outliers: 37% (moderately inflated)

benchmarking types/ByteArray/Sort32/2000/Block_W32
time                 1.007 ms   (989.5 μs .. 1.031 ms)
                     0.997 R²   (0.994 R² .. 0.999 R²)
mean                 995.4 μs   (985.1 μs .. 1.013 ms)
std dev              45.53 μs   (33.33 μs .. 64.35 μs)
variance introduced by outliers: 35% (moderately inflated)
```