@ciderale 2017-08-27
I'm not sure what happened with the failing 'GHC-8/OSX' build. The output does not tell much. Is that a problem of my commits or is it a general build problem?