@vincenthz 2017-08-28
@ciderale occasionally Travis build are not reliable. I've kicked it back, and hopefully will go through now