@vincenthz 2017-09-12
Ok I think that make sense, and we should go in this direction; although before proceeding I'll check further benchmarks / core dump to make sure there's no surprise