@ciderale 2017-09-12
Cool! Let's hope there is no suprise there. Let me know what your conclusions are. We then should also introduce a similar typeclass for immutable data, which covers even more of the algorithms.